requirement.txt
numpy
pandas
pyyaml
fastapi
uvicorn
polars
sklearn
